Typescriptify fake_api_client.js.
