+++ client/package.json
@@ -7,6 +7,7 @@
     "@reach/router": "^1.2.1",
     "@types/react": "^16.9.35",
     "@types/react-dom": "^16.9.8",
+    "@types/uuid": "^8.0.0",
     "bootstrap": "^3.3.7",
     "immutable": "^4.0.0-rc.9",
     "jquery": "^3.5.1",
